requirements.txt
json
kafka-python
